Extract inline <script> tags and move outside markdown area
remarkjs uses a <textarea> element for the slide markdown container, but this means that <script> tags inside the markdown can never work. (It's blocked at the browser level for obvious security reasons.)

This commit finds bare script tags in the markdown body and moves them to the end of the internal xaringan JS file. Care is taken to avoid <script> tags inside code chunks. <script> tags must appear at the start of a line and can be on a single line or span multiple lines.